Code refactorisation
